correct rapide contourh

+++ include/contour.h
@@ -1,15 +1,13 @@
 #ifndef CONTOUR_H
 #define CONTOUR_H
 
-#include "types_macros.h"
-
 typedef enum {
 	NORD, SUD, EST, OUEST
 } Orientation;
 
 typedef struct {
-	UINT x;
-	UINT y;
+	double x;
+	double y;
 	Orientation o;
 } RobotContour;
